% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps_data.R
\docType{data}
\name{aps_data}
\alias{aps_data}
\title{Data from APS study}
\format{
\subsection{\code{apsdata}}{

A data frame with 10,000 rows and 7 columns:
\describe{
\item{ID}{Subject identifier}
\item{x,z}{Covariates to adjust for}
\item{V}{Indicator of ever vaccinated while under follow-up}
\item{D_obs}{Day of observed vaccination (relative to d0)}
\item{Y}{Time to event in days (relative to d0)}
\item{event}{Event indicator, event = 0 if censored}
}
}
}
\source{
Created from code
}
\usage{
aps_data
}
\description{
Minimal data set mimicking data from an observational vaccine study.
Used to apply/test methods.
}
\keyword{datasets}
