% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{avecRentrez}
\alias{avecRentrez}
\title{function to query NCBI database servers using the "rentrex" library}
\usage{
avecRentrez(
  DB = "nucleotide",
  max.nr.recs = 20000,
  chunkSize = 100,
  accOnly = TRUE
)
}
\arguments{
\item{DB}{database}

\item{max.nr.recs}{maximun number of records to retrieve, there are limitations in the fns and server sides}

\item{chunkSize}{number of records to retrieve at once}

\item{accOnly}{boolean indicator for getting only accession codes or whole records}
}
\description{
function to query NCBI database servers using the "rentrex" library
}
\keyword{internal}
