% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{c19.fasta.data}
\alias{c19.fasta.data}
\title{function to obtain FASTA sequence of the SARS-CoV-2 virus}
\usage{
c19.fasta.data(src = "livedata")
}
\arguments{
\item{src}{argument to indicate where the data is being retrieved from}
}
\description{
function to obtain FASTA sequence of the SARS-CoV-2 virus
}
