% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_vaccination_testing.R
\name{c19.OWID.data}
\alias{c19.OWID.data}
\title{function to read data from OWID repos}
\usage{
c19.OWID.data(repo, disclaimer = TRUE)
}
\arguments{
\item{repo}{URL for reading the data}

\item{disclaimer}{indicate whether the information about the source of the data is disclosed}
}
\value{
datafram object
}
\description{
function to read data from OWID repos
}
