% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{c19.ptree.data}
\alias{c19.ptree.data}
\title{function to obtain "Tree of complete SARS-CoV-2 Sequences as obtained from NCBI"}
\usage{
c19.ptree.data(src = "livedata")
}
\arguments{
\item{src}{argument to indicate where the data is being retrieved from}
}
\description{
function to obtain "Tree of complete SARS-CoV-2 Sequences as obtained from NCBI"
}
