% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{calculate_ci}
\alias{calculate_ci}
\title{Calculate credible intervals for samples (e.g. output from Monte Carlo or Markov Chain Monte Carlo methods)}
\usage{
calculate_ci(samples, level, simple = TRUE)
}
\arguments{
\item{samples}{Vector of samples}

\item{level}{Credible level}

\item{simple}{If TRUE computation through quantiles, otherwise smallest interval containing the respective probability mass is computed}
}
\value{
Vector with upper and lower credible interval
}
\description{
Calculate credible intervals for samples (e.g. output from Monte Carlo or Markov Chain Monte Carlo methods)
}
\examples{
testsamples <- rnorm(1000,0,1)
level <- 0.95
simple <- TRUE
testci <- calculate_ci(testsamples, level,simple)
print(testci)

}
