% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{checkGeoLoc}
\alias{checkGeoLoc}
\title{function to check the geographical location}
\usage{
checkGeoLoc(data, geo.loc = NULL)
}
\arguments{
\item{data}{data.frame with data from covid19}

\item{geo.loc}{list of locations}
}
\value{
list of capitalized locations within the data
}
\description{
function to check the geographical location
}
\keyword{internal}
