% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{chk.TS.data}
\alias{chk.TS.data}
\title{auxiliary function to check whether a dataset is composed of time series data}
\usage{
chk.TS.data(data, xtp = FALSE)
}
\arguments{
\item{data}{data set to consider}

\item{xtp}{indicator whether to stop the program if data is not time series}
}
\value{
a boolean indicator whether the data contains time series values or not
}
\description{
auxiliary function to check whether a dataset is composed of time series data
}
\keyword{internal}
