% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_d.R
\name{compute_psi_d}
\alias{compute_psi_d}
\title{Compute day- and covariate- specific risks psi_v(d,x)}
\usage{
compute_psi_d(fit_0, fit_1, time_name, t0, tau, gp_list)
}
\arguments{
\item{fit_0}{A fitted survival model for the unvaccinated group}

\item{fit_1}{A fitted survival model for the vaccinated group}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{t0}{The time (relative to time of vaccination) at which to return
cumulative incidence estimates.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{gp_list}{List of marginalizing distributions. Used to determine which x, d
will be used for prediction}
}
\value{
A data frame containing predicted risk for each (d,x) of interest under
vaccine and no vaccine
}
\description{
Given the fitted survival models, compute estimates of psi_v(d,x)
for all d,x of interest.
}
