% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ci.R
\name{compute_wald_logodds_ci}
\alias{compute_wald_logodds_ci}
\alias{compute_wald_log_ci}
\title{Compute transformed Wald intervals for cumulative incidence/VE}
\usage{
compute_wald_logodds_ci(x, boot_x, z_crit_x)

compute_wald_log_ci(x, boot_x, z_crit_x)
}
\arguments{
\item{x}{Point estimate}

\item{boot_x}{Numeric vector containing bootstrapped point estimates}

\item{z_crit_x}{Critical value to use to construct Wald-style confidence interval}
}
\value{
Numeric vector containing the lower and upper confidence interval, and
standard error used to confidence interval
}
\description{
\code{compute_wald_logodds_ci}returns confidence interval based on transforming
Wald confidence intervals on the scale of \code{log(x/(1-x))}
\code{compute_wald_log_ci}returns confidence interval based on transforming
Wald confidence intervals on the scale of \code{log(1-x)}
}
