% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_Canada.R
\name{covid19.Canada.data}
\alias{covid19.Canada.data}
\title{function to import data for Canada
as reported by Health Canada
     https://health-infobase.canada.ca/src/data/covidLive/covid19.csv}
\usage{
covid19.Canada.data(
  data.fmt = "TS",
  local.data = FALSE,
  debrief = FALSE,
  acknowledge = FALSE
)
}
\arguments{
\item{data.fmt}{"TS" for TimeSeries of cumulative cases or "original" for the data as original reported}

\item{local.data}{boolean flag to indicate whether the data will be read from the local repo, in case of connectivity issues or data integrity}

\item{debrief}{boolean specifying whether information about the read data is going to be displayed in screen}

\item{acknowledge}{boolean flag to indicate that the user acknowledges where the data is coming from.  If FALSE, display data acquisition messages.}
}
\value{
a dataframe with the latest data reported by "OpenData Toronto" for the city of Toronto, ON - Canada
}
\description{
function to import data for Canada
as reported by Health Canada
     https://health-infobase.canada.ca/src/data/covidLive/covid19.csv
}
