% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{covid19.genomic.data}
\alias{covid19.genomic.data}
\title{main master (wrapper) function to obtain different types of genomic data for the SARS-CoV-2 virus}
\usage{
covid19.genomic.data(
  type = "genome",
  src = "livedata",
  graphics.ON = TRUE,
  accOnly = TRUE
)
}
\arguments{
\item{type}{type of data to retrieve, options are: 'genome', 'genomic', 'fasta', 'nucleotide', 'protein', 'ptree'}

\item{src}{source of the data: "livedata", "repo" or "local"}

\item{graphics.ON}{boolean option for display associated graphics}

\item{accOnly}{boolean indicator for getting only accession codes or whole records}
}
\description{
main master (wrapper) function to obtain different types of genomic data for the SARS-CoV-2 virus
}
