% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_vaccination_testing.R
\name{covid19.testing.data}
\alias{covid19.testing.data}
\title{function to read data related to covid19 Testing from OWID repo}
\usage{
covid19.testing.data(tgt = "testing", disclaimer = TRUE)
}
\arguments{
\item{tgt}{selects between time series data ('testing') and details and overall view of the data ("testing.details")}

\item{disclaimer}{indicates whether the information about the source of the data is disclosed}
}
\value{
dataframe containing list of countries (Entity) and testing data
}
\description{
function to read data related to covid19 Testing from OWID repo
}
