% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_dashboard.R
\name{covid19dashboard}
\alias{covid19dashboard}
\title{covid19.analytics explorer dashboard}
\usage{
covid19dashboard(locn = NULL)
}
\arguments{
\item{locn}{geographical location to use as default}
}
\value{
list with shinyApp UI and server
}
\description{
covid19.analytics explorer dashboard
}
