% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_dashboard-wrapper.R
\name{covid19Explorer}
\alias{covid19Explorer}
\title{covid19.analytics explorer dashboard}
\usage{
covid19Explorer(locn = NULL)
}
\arguments{
\item{locn}{geographical location to use as default}
}
\description{
covid19.analytics explorer dashboard
}
