% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfinxMain.R
\docType{data}
\name{DataInputExampleFile}
\alias{DataInputExampleFile}
\title{The TIP49 dataset of protein interactions (AP-MS).}
\format{A matrix with 1581 rows (proteins) and 70 variables (projects).}
\source{
M.E. Sardiu, Y. Cai, J. Jin, S.K. Swanson, R.C. Conaway, J.W.
Conaway, L. Florens, M.P. Washburn Probabilistic assembly of human protein
interaction networks from label-free quantitative proteomics. Proc. Natl.
Acad. Sci. USA, 105 (2008), pp. 1454 to 1459.
}
\usage{
DataInputExampleFile
}
\description{
A strictly numeric input matrix with unique proteins as rownames
and unique projects as colnames. The cells of the matrix are filled with
the associated \strong{peptide counts}. Cells that have no associated
peptide counts are filled with a zero. This specific example dataset is
derived from the publication of Sardiu et al. (see below). It contains
complexes involved in chromatin remodeling and consists of 35 bait-specific
projects and 35 negative controls.
}
\keyword{datasets}

