\name{db}
\alias{db}
\title{db}
\description{
  This function is not used. It provides an example of the dualbrothers function.
}
\usage{db(x, ...)}
\arguments{
  \item{x}{A string in quotes that requires the x.phy file to be in the current working directory.}
  \item{...}{optional arguments}
}
\details{
 See help(dualbrothers) for details.
}
\value{a dualbrothers object}
