% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_matching_ve.R
\name{estimate_marginal_risk}
\alias{estimate_marginal_risk}
\title{Get the marginal cumulative incidence in the treated and untreated
groups based on a Cox model.}
\usage{
estimate_marginal_risk(
  adata,
  adata_outcome_name,
  adata_event_name,
  adata_trt_name,
  adjust,
  times,
  censor_time,
  tau,
  separate = FALSE,
  return_models = TRUE
)
}
\arguments{
\item{adata}{A data fram that represents the analysis data set of a clinical trial.}

\item{adata_outcome_name}{Character string specifying the time to event variable in \code{adata}. The
time should be the time to event from vaccination/matched index date}

\item{adata_event_name}{Character string specifying the event variable in \code{adata}}

\item{adata_trt_name}{Character string specifying the treatment variable in \code{adata}}

\item{adjust}{A character vector  of variables in \code{adata} to adjust for, or a formula object.
Do not include treatment variable.}

\item{times}{A numeric vector of the time points at which to return cumulative incidence estimates}

\item{censor_time}{Time at which observations should be censored. Typically the \code{max(times)}.}

\item{tau}{waiting period}

\item{separate}{If \code{TRUE}, models are fit in the treated and untreated groups separately}

\item{return_models}{If \code{TRUE}, fitted survival mdoels are returned.}
}
\value{
A list containing the following:
\describe{
\item{estimates}{A list with three named elements, "risk_0", "risk_1", "ve". Each
element is a vector containing the estimates for the given term at each time point in \code{times}}
\item{models}{If \code{return_models = TRUE}, the models used to compute risk and VE}
}
}
\description{
This function uses a Cox model to predict the marginal cumulative
incidence of an endpoint \code{t0} days after vaccination. The marginal
cumulative incidence is computed by predicting individual survival probabilities
and averaging these predictions over everyone in the treatment group.
}
