% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsve.R
\name{estimates_to_df}
\alias{estimates_to_df}
\title{Convert estimates matrix to data frame}
\usage{
estimates_to_df(estimates)
}
\arguments{
\item{estimates}{A matrix containing the point estimates and lower and
upper confidence bounds}
}
\value{
A data frame of the estimates
}
\description{
Convert estimates matrix to data frame
}
