\name{extractbasename}
\alias{extractbasename}
\title{extract the basename from an rbrothers object}
\description{
  This function returns the basename in an rbrothers object
}
\usage{extractbasename(x)}
\arguments{
  \item{x}{an rbrothers object}
}
\value{a string (also attainable via x$basename where c is the rbrothers object)}
