% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_models.R
\name{gen.glm.model}
\alias{gen.glm.model}
\title{function to generate models using GLM}
\usage{
gen.glm.model(y, family = Gamma(link = "log"))
}
\arguments{
\item{y}{vector containing the data to fit}

\item{family}{family to use in the GLM method}
}
\description{
function to generate models using GLM
}
\keyword{internal}
