% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_bar_chart.R
\name{gender_bar_chart}
\alias{gender_bar_chart}
\title{Function to plot the barchart of the total number by gender}
\usage{
gender_bar_chart(data_df, x_title, y_title, label_name)
}
\arguments{
\item{data_df, }{dataframe generated from the function total_df}

\item{x_title, }{title for x axis.}

\item{y_title, }{title for y axis}

\item{label_name, }{label for the total number of male and female}
}
\value{
a barchart
}
\description{
Function to plot the barchart of the total number by gender
}
\examples{
\dontrun{
totalbarchart <- gender_bar_chart(data_df)
}
}
