% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_models.R
\name{genModel}
\alias{genModel}
\title{function to generate models using Linear Regression "LM"}
\usage{
genModel(y, deg = 1)
}
\arguments{
\item{y}{vector containing the data to fit}

\item{deg}{degree of the polynomial fit}
}
\description{
function to generate models using Linear Regression "LM"
}
\keyword{internal}
