% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_geos.R
\name{geographicalRegions}
\alias{geographicalRegions}
\title{function to define continents and its constituent countries}
\usage{
geographicalRegions(cont = NULL)
}
\arguments{
\item{cont}{optional argumetn, to specify a particular continent; if no argument is given then it returns all the continents and countries for each}
}
\value{
list with the composition of continents
}
\description{
function to define continents and its constituent countries
}
