% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityD-class.R
\docType{methods}
\name{getP,purityD-method}
\alias{getP,purityD-method}
\alias{getP}
\title{Get peaklist for a purityD object}
\usage{
\S4method{getP}{purityD}(x)
}
\arguments{
\item{x}{object; purityD object}
}
\value{
peaks
}
\description{
output peak list for a purityD object
}
\examples{
datapth <- system.file("extdata", "dims", "mzML", package="msPurityData")
inDF <- Getfiles(datapth, pattern=".mzML", check = FALSE, cStrt = FALSE)
ppDIMS <- purityD(fileList=inDF, cores=1, mzML=TRUE)
peaks <- getP(ppDIMS)
}
