% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_checks.R
\name{integrity.check}
\alias{integrity.check}
\title{function that determines whether there are integrity issues within the datasets
or changes to the structure of the data as reported by JHU/CCSEGIS}
\usage{
integrity.check(data, datasetName = "", recommend = TRUE)
}
\arguments{
\item{data}{dataset to analyze}

\item{datasetName}{optional argument to display the name of the dataset}

\item{recommend}{optional flag to recommend further actions}
}
\description{
function that determines whether there are integrity issues within the datasets
or changes to the structure of the data as reported by JHU/CCSEGIS
}
