% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{lastyear.date}
\alias{lastyear.date}
\title{function that returns the date from one year ago}
\usage{
lastyear.date()
}
\description{
function that returns the date from one year ago
}
\keyword{internal}
