% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{loadLibrary}
\alias{loadLibrary}
\title{function to check and load an specific set of libraries}
\usage{
loadLibrary(lib)
}
\arguments{
\item{lib}{is a list of packages to be loaded}
}
\description{
function to check and load an specific set of libraries
}
\keyword{internal}
