% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_plts.R
\name{log.sc.setup}
\alias{log.sc.setup}
\title{aux fn to create a log-scale pulldown option in plotly plots}
\usage{
\method{log}{sc.setup}(nbr.traces)
}
\arguments{
\item{nbr.traces}{number of traces in order to set Ts/Fs}
}
\description{
aux fn to create a log-scale pulldown option in plotly plots
}
\keyword{internal}
