% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matched_adata.R
\name{match_adata}
\alias{match_adata}
\title{Perform matching and get matched adata back}
\usage{
match_adata(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  id_name,
  matching_vars,
  replace = FALSE,
  seed = NULL,
  tau,
  pair_censoring = TRUE
)
}
\arguments{
\item{data}{A data frame for which matches are sought}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string representing the original event variable in \code{matched_data}}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{id_name}{Character string specifying name of a unique individual
identifier in \code{data}.}

\item{matching_vars}{A character vector containing the names of variables in
\code{data} to match on .}

\item{replace}{Logical: Should matching be done with replacement?}

\item{seed}{Numeric: seed to set prior to performing random matching}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}
}
\value{
A list containing the matched analysis dataset and the matched cohort
}
\description{
Perform matching and get matched adata back
}
