% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_rolling_cohort.R
\name{match_rolling_cohort}
\alias{match_rolling_cohort}
\title{Perform exact 1:1 matching using "rolling cohort" design}
\usage{
match_rolling_cohort(
  data,
  outcome_name,
  trt_name,
  time_name,
  id_name,
  matching_vars,
  replace = FALSE,
  seed = NULL
)
}
\arguments{
\item{data}{A data frame for which matches are sought}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{id_name}{Character string specifying name of a unique individual
identifier in \code{data}.}

\item{matching_vars}{A character vector containing the names of variables in
\code{data} to match on .}

\item{replace}{Logical: Should matching be done with replacement?}

\item{seed}{Numeric: seed to set prior to performing random matching}
}
\value{
A list containing the following:
\describe{
\item{matched_data}{A data frame containing the matched individuals}
\item{n_unmatched_cases}{The number of vaccinated individuals for whom no match was found}
\item{discarded}{A logical vector denoting whether individual in the source population was
not included in matched_cohort}
}
}
\description{
This function creates a matched cohort based on the rolling cohort design.
Individuals are eligible to be selected as unvaccinated "controls" only if they
are unvaccinated and endpoint free on the potential matching date
}
