% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{normalize}
\alias{normalize}
\title{Center and rescale irregular time series (NA values are removed from computing means / standard deviations)}
\usage{
normalize(xin, center = TRUE, scale = TRUE)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{center}{logical center to zero mean?}

\item{scale}{logical normalize to unit standard deviation?}
}
\value{
Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate
}
\description{
Center and rescale irregular time series (NA values are removed from computing means / standard deviations)
}
\examples{
# Load Monticchio example data from PTBoxProxydata
library(PTBoxProxydata)
mng <- PTBoxProxydata::ProxyDataManager()
monticchiodata <- PTBoxProxydata::load_set(mng,'monticchio_testset',zoo_format = 'zoo', force_file=TRUE)
# Normalize proxy data in proxytibble
monticchiodata_normalized <- normalize(monticchiodata)

# Normalize single zoo
monticchiozoo_normalized <- normalize(monticchiodata$proxy_data[[1]])
# Plot zoo data
plot(monticchiozoo_normalized)
plot(monticchiodata_normalized$proxy_data[[1]])

}
