% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsve-package.R
\docType{package}
\name{obsve-package}
\alias{obsve-package}
\title{obsve: Causal Estimand for Observational Vaccine Studies}
\description{
Functions to estimate vaccine efficacy in observational vaccine studies.
}
\author{
\strong{Maintainer}: Emily Wu \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
