% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{outputType}
\alias{outputType}
\title{function to validate the type of file format and open the corresponding file}
\usage{
outputType(
  device,
  dirSave = NULL,
  fileName,
  fileFmts,
  combinePlts,
  msg = "plot"
)
}
\arguments{
\item{device}{string to select the output format: 'PDF'/'PNG'/'JPEG' or 'screen'}

\item{dirSave}{specify a valid directory where to save the file}

\item{fileName}{name of the file}

\item{fileFmts}{list with the acceptable format types, eg. c("PDF","PNG","JPEG")}

\item{combinePlts}{boolean flag indicating whether plots are combined or not}

\item{msg}{optional string to customiza message to the user}
}
\description{
function to validate the type of file format and open the corresponding file
}
\keyword{internal}
