% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{paleodata_windowing}
\alias{paleodata_windowing}
\title{Extract all samples in a given time period from an irregular time series}
\usage{
paleodata_windowing(xin, start_date, end_date)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{start_date}{Lower end of time window (included in the output `zoo::zoo`)}

\item{end_date}{Upper end of time window (included in the output `zoo::zoo`)}
}
\value{
Proxytibble with proxy data in `zoo::zoo` format or irregular time series object (`zoo::zoo`) containing only the samples in the defined window
}
\description{
Extract all samples in a given time period from an irregular time series
}
\examples{
# Load Monticchio example data from PTBoxProxydata
library(PTBoxProxydata)
mng <- PTBoxProxydata::ProxyDataManager()
monticchiodata <- PTBoxProxydata::load_set(mng,'monticchio_testset',zoo_format = 'zoo', force_file=TRUE)
# Reduce to samples between 30ka BP and 60ka BP
monticchiodata <- paleodata_windowing(monticchiodata,30000,60000)
# Extract zoo with samples from 30ka BP to 40ka BP
monticchiozoo <- paleodata_windowing(monticchiodata$proxy_data[[1]],30000,40000)
# Plot zoo data
plot(monticchiozoo)

}
