% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandemics.R
\name{pandemics.data}
\alias{pandemics.data}
\title{function to retrieve historical pandemics data}
\usage{
pandemics.data(acknowledge = TRUE, show = FALSE, tgt = "pandemics")
}
\arguments{
\item{acknowledge}{displays details on the data sources}

\item{show}{displays data}

\item{tgt}{which data set to read -- options are  'pandemics' OR 'pandemics_vaccines'}
}
\value{
data.frame
}
\description{
function to retrieve historical pandemics data
}
