% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandemics.R
\name{pandemics.loaddata}
\alias{pandemics.loaddata}
\title{internal function to retrieve historical data on pandemics}
\usage{
pandemics.loaddata(
  tgt.file = "pandemics.RDS",
  acknowledge = TRUE,
  show = FALSE,
  src.descr = ""
)
}
\arguments{
\item{tgt.file}{which data set to read}

\item{acknowledge}{displays details on the data sources}

\item{show}{displays data}

\item{src.descr}{description of the source of the data}
}
\value{
data.frame
}
\description{
internal function to retrieve historical data on pandemics
}
