% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timepoints.R
\name{plot_timepoints}
\alias{plot_timepoints}
\title{Plot cumulative incidence/VE estimates}
\usage{
plot_timepoints(object)
}
\arguments{
\item{object}{Object created by timepoints}
}
\value{
ggplot
}
\description{
Plot cumulative incidence/VE estimates
}
