% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_print.R
\name{print}
\alias{print}
\title{method associated with 'report' objects}
\usage{
print(report.obj)
}
\arguments{
\item{report.obj}{'report' object}
}
\description{
method associated with 'report' objects
}
\keyword{internal}
