% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_print.R
\name{print.report}
\alias{print.report}
\title{function associated to the 'report' object method}
\usage{
\method{print}{report}(report.obj)
}
\arguments{
\item{report.obj}{'report' object}
}
\description{
function associated to the 'report' object method
}
\keyword{internal}
