% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psdata_cl.R
\docType{data}
\name{psdata_cl}
\alias{psdata_cl}
\title{Simulated dataset for PSweight}
\format{
A data frame with 1500 rows and 9 columns.
}
\usage{
data(psdata_cl)
}
\description{
This is a simulated observational study with three treatment groups to illustated the ulity of PSweight.
}
\details{
The simulated dataset includes 1500 rows, with each row represents information recorded from each individual.
There are 9 variables (columns). The treatment is the variable trt, which has two treatment arms. The clt is the cluster level.
The outcome of interest is variable Y. cov1-cov6 are pre-treatment covariates among which cov1-cov5 are continous, and cov6 is binary.
}
\examples{
# data("psdata_cl")
}
\keyword{dataset}
