% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{red.devel.ver}
\alias{red.devel.ver}
\title{function to redirect users to install devel version from github}
\usage{
red.devel.ver(fileRDS, force = TRUE)
}
\arguments{
\item{fileRDS}{missing data file}

\item{force}{boolean flag to force stoping the code}
}
\description{
function to redirect users to install devel version from github
}
