% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_anal.R
\name{report.summary}
\alias{report.summary}
\title{function to summarize the current situation, will download the latest data and summarize the top provinces/cities per case}
\usage{
report.summary(
  cases.to.process = "ALL",
  Nentries = 10,
  graphical.output = TRUE,
  saveReport = FALSE
)
}
\arguments{
\item{cases.to.process}{which data to process: "TS" --time series--, "AGG" --aggregated-- or "ALL" --time series and aggregated--}

\item{Nentries}{number of top cases to display}

\item{graphical.output}{flag to deactivate graphical output}

\item{saveReport}{flag to indicate whether the report should be saved in a file}
}
\description{
function to summarize the current situation, will download the latest data and summarize the top provinces/cities per case
}
\examples{
# displaying top 10s
report.summary()

# get the top 20
report.summary(20)

}
