% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadsStats-utils.R
\name{retrievePckgData}
\alias{retrievePckgData}
\title{function to download the data from the CRAN logs for an specific package}
\usage{
retrievePckgData(pckg = NULL, t0 = lastyear.date(), t1 = today())
}
\arguments{
\item{pckg}{is the name of the package to look for the downloads data}

\item{t0}{is the initial date}

\item{t1}{is the final date}
}
\value{
a list composed of the stats from the original time frame and the last month
}
\description{
function to download the data from the CRAN logs for an specific package
}
\examples{
\donttest{
retrievePckgData("ehelp")
retrievePckgData("ehelp","2018-01-01","2020-01-01")
}
}
