% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{select.per.loc}
\alias{select.per.loc}
\title{auxiliary function to select data based on geographical location}
\usage{
select.per.loc(data, geo.loc)
}
\arguments{
\item{data}{data set to process}

\item{geo.loc}{geographical location, can be a country, region, province or city}
}
\description{
auxiliary function to select data based on geographical location
}
\keyword{internal}
