% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityX-class.R
\docType{methods}
\name{show,purityX-method}
\alias{show,purityX-method}
\title{Show method for purityX}
\usage{
\S4method{show}{purityX}(object)
}
\arguments{
\item{object}{object; purityX object}
}
\value{
a print statement of regarding object
}
\description{
Show method for purityX object
}
