% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated dataset}
\format{
\subsection{\code{simdata}}{

A data frame with 10,000 rows and 7 columns:
\describe{
\item{ID}{Subject identifier}
\item{x1,x2}{Covariates to adjust for}
\item{V}{Indicator of ever vaccinated while under follow-up}
\item{D_obs}{Day of observed vaccination (relative to d0)}
\item{Y}{Time to event in days (relative to d0)}
\item{event}{Event indicator, event = 0 if censored}
}
}
}
\source{
Created from code
}
\usage{
simdata
}
\description{
Minimal data set mimicking data from an observational vaccine study.
Used to apply/test methods.
}
\keyword{datasets}
