% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_anal.R
\name{single.trend}
\alias{single.trend}
\title{function to visualize different indicators for trends in daily changes of cases reported as time series data}
\usage{
single.trend(ts.data, confBnd = TRUE, info = "")
}
\arguments{
\item{ts.data}{time series data}

\item{confBnd}{optional argument to remove the drawing of a confidence band}

\item{info}{addtional information to display in plots}
}
\description{
function to visualize different indicators for trends in daily changes of cases reported as time series data
}
\examples{
\donttest{
tor.data <- covid19.Toronto.data()
single.trend(tor.data[tor.data$status=="Active Cases",]) 
}

ts.data <- covid19.data("ts-confirmed")
ont.data <- ts.data[ ts.data$Province.State == "Ontario",]
single.trend(ont.data)

single.trend(ts.data[ ts.data$Country.Region=="Italy",])
}
