% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{time.intervals}
\alias{time.intervals}
\title{function to identify units of times and time intervals for customizing axes of plots}
\usage{
\method{time}{intervals}(tot.days)
}
\arguments{
\item{tot.days}{number of days in time interval to consider}
}
\value{
a list with the unit of time and time interval
}
\description{
function to identify units of times and time intervals for customizing axes of plots
}
\keyword{internal}
