% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{time.units}
\alias{time.units}
\title{function to identify units of times based on the number of days}
\usage{
\method{time}{units}(tot.days)
}
\arguments{
\item{tot.days}{number of days}
}
\value{
a vector with the "time units" included in the total number of days
}
\description{
function to identify units of times based on the number of days
}
\keyword{internal}
