% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_plts.R
\name{totals.plt}
\alias{totals.plt}
\title{function to plot total number of cases per day for different groups}
\usage{
totals.plt(
  data0 = NULL,
  geo.loc = NULL,
  interactive.fig = TRUE,
  fileName = NULL
)
}
\arguments{
\item{data0}{time series dataset to process, default all the possible cases: 'confirmed' and 'deaths' for all countries/regions}

\item{geo.loc}{geographical location, country/region or province/state to restrict the analysis to}

\item{interactive.fig}{swith to turn off/on an interactive plot}

\item{fileName}{file where to save the HTML version of the interactive figure}
}
\description{
function to plot total number of cases per day for different groups
}
\examples{
# retrieve time series data
TS.data <- covid19.data("ts-ALL")

# static and interactive plot 
totals.plt(TS.data)

}
