% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_PSweight.R
\name{vcov.PSweight}
\alias{vcov.PSweight}
\title{Variance of PSweight}
\usage{
\method{vcov}{PSweight}(object, ...)
}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{vcov}}
}
\description{
The \code{\link{vcov}} method for class "PSweight"
}
