% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{loadLibrary}
\alias{loadLibrary}
\title{auxiliary function to check and load an specific set of libraries}
\usage{
loadLibrary(lib)
}
\arguments{
\item{lib}{is a list of packages to be loaded}
}
\description{
auxiliary function to check and load an specific set of libraries
}
\keyword{internal}
